refactor add_commands() out into commands module
