+++ lib/svg/__init__.py
@@ -1,4 +1,5 @@
-from .svg import color_block_to_point_lists, render_stitch_plan
-from .units import *
-from .path import apply_transforms, get_node_transform, get_correction_transform, line_strings_to_csp, point_lists_to_csp
 from .guides import get_guides
+from .path import apply_transforms, get_node_transform, get_correction_transform, line_strings_to_csp, point_lists_to_csp
+from .rendering import color_block_to_point_lists, render_stitch_plan
+from .svg import get_document, generate_unique_id
+from .units import *
